5758 - extract validateLink to its own file
